requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic[email]==2.9.2
pyjwt==2.9.0
bcrypt==4.2.0
python-multipart==0.0.12
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
redis==5.2.0
pytest==8.3.0
httpx==0.27.0
email-validator==2.2.0
